Unify the Autonomi REST client core in autvid_common::antd
AntdClient owns the shared plumbing both services duplicated: bearer
auth, circuit breaking, jittered exponential retry loops, streaming
file upload with SHA-256 integrity, and metrics via the new
AntdMetricsRecorder trait (NoopRecorder default). DTOs move to
antd::types; raw_endpoint_unavailable and
is_missing_file_upload_endpoint classifiers move alongside.

Services keep thin wrappers so observable behavior is unchanged:
- rust_stream: 3-attempt latency-sensitive retries (250ms base),
  raw-endpoint fallback to base64 JSON, no metrics (Noop recorder)
- rust_admin: single-shot instrumented requests, 5-attempt cost
  estimate loop, slow-backoff file upload retries; AdminMetrics
  implements the recorder

Both mock-antd test suites pass unchanged, preserving metric names the
Grafana dashboards depend on.
