+++ crates/common/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod antd;
 pub mod env;
 pub mod error;
 pub mod health;
